connecting backend to frontend version 1.0

+++ AndroidWithKotlin/WeatherApp/app/src/main/java/com/example/weatherapp/models/Sys.kt
@@ -1,11 +1,9 @@
 package com.example.weatherapp.models
 
-import android.app.Notification.MessagingStyle.Message
-
 data class Sys(
     val type: Int,
     val message: Double,
     val country: String,
-    val sunrise: Int,
-    val sunset: Int
+    val sunrise: Long,
+    val sunset: Long
 ) : java.io.Serializable
